requirements.txt
Django==5.1.4
djangorestframework==3.15.2
drf-yasg==1.21.8
pytest==8.3.4
pytest-django==4.9.0